Removed screenshots for responsive preview and moved integration tests to single file
